requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx>=0.24.0
pytest>=7.0.0

google-generativeai>=0.7.0
supabase
pydantic>=2.0.0
